refactor: EventPanelButtons 클래스 책임 분리
- Skip, Auto버튼이 Event Panel에 종속되지 않고 독립적으로 실행되는 경우가 존재하여 책임을 분리
