+++ Assets/Min/Scripts/UIs/AutoAndSkipButtonPainter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9d6ba22c4230de1428e0cc4e98e98d41
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
